Add a functional tests folder, with Subscription and SubscriptionItem tests from StripeTests
