+++ src/test/java/com/stripe/StripeTest.java
@@ -165,31 +165,6 @@ public class StripeTest {
 		return params;
 	}
 
-	static Subscription createDefaultSubscription(Customer customer)
-			throws StripeException {
-		Plan plan = Plan.create(getUniquePlanParams());
-		Map<String, Object> subCreateParams = new HashMap<String, Object>();
-		HashMap<String, Object> items = new HashMap<String, Object>();
-		HashMap<String, Object> item = new HashMap<String, Object>();
-
-		item.put("plan", plan.getId());
-		item.put("quantity", 1);
-		items.put("0", item);
-
-		subCreateParams.put("items", items);
-		subCreateParams.put("customer", customer.getId());
-		return Subscription.create(subCreateParams);
-	}
-
-	static SubscriptionItem createDefaultSubscriptionItem(Subscription subscription)
-			throws StripeException {
-		Plan plan = Plan.create(getUniquePlanParams());
-		Map<String, Object> subItemParams = new HashMap<String, Object>();
-		subItemParams.put("plan", plan.getId());
-		subItemParams.put("subscription", subscription.getId());
-		return SubscriptionItem.create(subItemParams);
-	}
-
 	static InvoiceItem createDefaultInvoiceItem(Customer customer)
 			throws StripeException {
 		Map<String, Object> invoiceItemParams = new HashMap<String, Object>();
@@ -205,13 +180,6 @@ public class StripeTest {
 		customerWithPlanParams.putAll(defaultCustomerParams);
 		customerWithPlanParams.put("plan", plan.getId());
 		return Customer.create(customerWithPlanParams);
-	}
-
-	static Recipient createDefaultRecipient()
-			throws StripeException {
-		Map<String, Object> recipientParams = new HashMap<String, Object>();
-		recipientParams.putAll(defaultRecipientParams);
-		return Recipient.create(recipientParams);
 	}
 
 	static Map<String, Object> getSubscriptionParams() throws StripeException {
@@ -453,195 +421,6 @@ public class StripeTest {
 		listParams.put("count", 1);
 		List<Charge> charges = Charge.all(listParams).getData();
 		assertEquals(charges.size(), 1);
-	}
-
-	// Subscription Tests:
-	@Test
-	public void testUpdateSubscription() throws StripeException {
-		Plan plan = Plan.create(getUniquePlanParams());
-		Customer customer = Customer.create(defaultCustomerParams);
-		Map<String, Object> subscriptionParams = new HashMap<String, Object>();
-		subscriptionParams.put("plan", plan.getId());
-		Subscription sub = customer.updateSubscription(subscriptionParams);
-		assertEquals(sub.getPlan().getId(), plan.getId());
-		assertEquals(sub.getCustomer(), customer.getId());
-	}
-
-	@Test
-	public void testCancelSubscription() throws StripeException {
-		Plan plan = Plan.create(getUniquePlanParams());
-		Customer customer = createDefaultCustomerWithPlan(plan);
-		assertEquals(customer.getSubscriptions().getData().get(0).getStatus(), "active");
-		Subscription canceledSubscription = customer.cancelSubscription();
-		assertEquals(canceledSubscription.getStatus(), "canceled");
-	}
-
-	@Test
-	public void testCancelSubscriptionAtPeriodEnd() throws StripeException {
-		Plan plan = Plan.create(getUniquePlanParams());
-		Customer customer = createDefaultCustomerWithPlan(plan);
-		assertEquals(customer.getSubscriptions().getData().get(0).getStatus(), "active");
-		Map<String, Object> cancelParams = new HashMap<String, Object>();
-		cancelParams.put("at_period_end", true);
-		Subscription canceledSubscription = customer
-				.cancelSubscription(cancelParams);
-		assertEquals(canceledSubscription.getStatus(), "active");
-		assertEquals(canceledSubscription.getCancelAtPeriodEnd(), true);
-	}
-
-	@Test
-	public void testNewStyleSubscriptionAPI() throws StripeException {
-		Plan plan = Plan.create(getUniquePlanParams());
-		Plan plan2 = Plan.create(getUniquePlanParams());
-		Customer customer = Customer.create(defaultCustomerParams);
-
-		// Create
-		Map<String, Object> subCreateParams = new HashMap<String, Object>();
-		subCreateParams.put("plan", plan.getId());
-		Subscription sub = customer.createSubscription(subCreateParams);
-		assertEquals(plan.getId(), sub.getPlan().getId());
-		customer = Customer.retrieve(customer.getId());
-		assertEquals(1, customer.getSubscriptions().getData().size());
-		assertEquals(sub.getId(), customer.getSubscriptions().getData().get(0).getId());
-
-		// Retrieve
-		Subscription retrievedSub = customer.getSubscriptions().retrieve(sub.getId());
-		assertEquals(sub.getId(), retrievedSub.getId());
-
-		// List
-		CustomerSubscriptionCollection list = customer.getSubscriptions().all(null);
-		assertEquals(1, list.getData().size());
-		assertEquals(sub.getId(), list.getData().get(0).getId());
-
-		// Update
-		Map<String, Object> subUpdateParams = new HashMap<String, Object>();
-		subUpdateParams.put("plan", plan2.getId());
-		sub = sub.update(subUpdateParams);
-		assertEquals(plan2.getId(), sub.getPlan().getId());
-
-		// Cancel
-		sub = sub.cancel(null);
-		assertNotNull(sub.getCanceledAt());
-	}
-
-	@Test
-	public void testTopLevelSubscriptionAPI() throws StripeException {
-		Plan plan = Plan.create(getUniquePlanParams());
-		Plan plan2 = Plan.create(getUniquePlanParams());
-		Customer customer = Customer.create(defaultCustomerParams);
-
-		// Create
-		Map<String, Object> subCreateParams = new HashMap<String, Object>();
-		subCreateParams.put("plan", plan.getId());
-		subCreateParams.put("customer", customer.getId());
-		Subscription sub = Subscription.create(subCreateParams);
-		assertEquals(plan.getId(), sub.getPlan().getId());
-		assertEquals(customer.getId(), sub.getCustomer());
-
-		customer = Customer.retrieve(customer.getId());
-		assertEquals(1, customer.getSubscriptions().getData().size());
-		assertEquals(sub.getId(), customer.getSubscriptions().getData().get(0).getId());
-
-		// Retrieve
-		Subscription retrievedSub = Subscription.retrieve(sub.getId());
-		assertEquals(sub.getId(), retrievedSub.getId());
-
-		// List
-		Map<String, Object> subAllParams = new HashMap<String, Object>();
-		subAllParams.put("plan", plan.getId());
-		subAllParams.put("customer", customer.getId());
-		SubscriptionCollection list = Subscription.all(subAllParams);
-		assertEquals(1, list.getData().size());
-		assertEquals(sub.getId(), list.getData().get(0).getId());
-		assertEquals(customer.getId(), list.getData().get(0).getCustomer());
-		assertEquals(plan.getId(), list.getData().get(0).getPlan().getId());
-
-		// Update
-		Map<String, Object> subUpdateParams = new HashMap<String, Object>();
-		subUpdateParams.put("plan", plan2.getId());
-		sub = sub.update(subUpdateParams);
-		assertEquals(plan2.getId(), sub.getPlan().getId());
-
-		// Cancel
-		sub = sub.cancel(null);
-		assertNotNull(sub.getCanceledAt());
-	}
-
-
-	@Test
-	public void testCreateSubscriptionThroughCollection() throws StripeException {
-		Plan plan = Plan.create(getUniquePlanParams());
-		Customer customer = Customer.create(defaultCustomerParams);
-
-		// Create
-		Map<String, Object> subCreateParams = new HashMap<String, Object>();
-		subCreateParams.put("plan", plan.getId());
-
-		Subscription sub = customer.getSubscriptions().create(subCreateParams);
-		assertEquals(plan.getId(), sub.getPlan().getId());
-
-		// Verify
-		customer = Customer.retrieve(customer.getId());
-		assertEquals(1, customer.getSubscriptions().getData().size());
-		assertEquals(sub.getId(), customer.getSubscriptions().getData().get(0).getId());
-	}
-
-	@Test
-	public void testSubscriptionItemCreate() throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams);
-		Subscription subscription = createDefaultSubscription(customer);
-		SubscriptionItem subscriptionItem = createDefaultSubscriptionItem(subscription);
-
-		assertEquals(subscriptionItem.getPlan().getName(), "J Bindings Plan");
-	}
-
-	@Test
-	public void testSubscriptionItemRetrieve() throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams);
-		Subscription subscription = createDefaultSubscription(customer);
-		SubscriptionItem subscriptionItem = createDefaultSubscriptionItem(subscription);
-
-		SubscriptionItem retrievedSubscriptionItem = SubscriptionItem
-			.retrieve(subscriptionItem.getId());
-		assertEquals(subscriptionItem.getId(), retrievedSubscriptionItem.getId());
-	}
-
-	@Test
-	public void testSubscriptionItemList() throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams);
-		Subscription subscription = createDefaultSubscription(customer);
-		SubscriptionItem subscriptionItem = createDefaultSubscriptionItem(subscription);
-
-		Map<String, Object> listParams = new HashMap<String, Object>();
-		listParams.put("subscription", subscription.getId());
-		SubscriptionItemCollection subscriptionItems = SubscriptionItem.list(listParams);
-		List<SubscriptionItem> subscriptionItemsData = subscriptionItems.getData();
-		assertEquals(subscriptionItemsData.size(), 2);
-	}
-
-	@Test
-	public void testSubscriptionItemUpdate() throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams);
-		Subscription subscription = createDefaultSubscription(customer);
-		SubscriptionItem subscriptionItem = createDefaultSubscriptionItem(subscription);
-
-		Map<String, Object> updateParams = new HashMap<String, Object>();
-		updateParams.put("quantity", 4);
-
-		SubscriptionItem updatedSubscriptionItem =
-			subscriptionItem.update(updateParams);
-		assertTrue(updatedSubscriptionItem.getQuantity() == 4);
-	}
-
-	@Test
-	public void testSubscriptionItemDelete() throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams);
-		Subscription subscription = createDefaultSubscription(customer);
-		SubscriptionItem subscriptionItem = createDefaultSubscriptionItem(subscription);
-
-		DeletedSubscriptionItem deletedSubscriptionItem = subscriptionItem.delete();
-		assertTrue(deletedSubscriptionItem.getDeleted());
-		assertEquals(deletedSubscriptionItem.getId(), subscriptionItem.getId());
 	}
 
 	// Invoice Tests:
