+++ src/test/java/com/stripe/model/SubscriptionItemTest.java
@@ -10,8 +10,13 @@ import org.junit.Before;
 import junit.framework.Assert;
 
 import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
 import org.junit.Test;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.*;
 
 public class SubscriptionItemTest extends BaseStripeTest {
